Initial commit with a quick, dirty and crude version of the project.
The app is just a copy with minimal changes of the code-metrics app.
Some error checking needs to be applied.
The input code should be compiled with gcc and errors reported.
=> Not a great deal.

The transpiler is the original version made by Ricardo Martín. Some work
on it remains to be done. Pascal looks right, but the other languages have
some glitches or bugs. Where is the plib? Format of errors is needed.
=> Some work remains here.